REFACTOR : 음식 리스트 받아오기 기능 수정
DTO 패키지 구조 변경
FoodListResponseDto 생성
FoodController getFoodList() return type 수정
